# Conflicts:
#	app/src/main/java/com/example/myapplication/ui/Principal/PrincipalScreen.kt
#	app/src/main/java/com/example/myapplication/ui/VistasObras/VistaObrasViewModel.kt

+++ app/src/main/java/com/example/myapplication/ui/VistasObras/VistaObrasViewModel.kt
@@ -14,8 +14,11 @@ class VistaObrasViewModel @Inject constructor(): ViewModel() {
     var uiState: MutableStateFlow<VistaObrasState> = _uiState
     fun getObra(id: Int): Obra? {
         //ir al repo y filtrar
-        val obra = ProveedorObras.obras.find { it.obraId == id }
-        return obra
+        val obra = ProveedorObras.obras.find { it.obraId.toInt() == id }
+
+            return obra
+
+
     }
     fun updateObra(obra: Obra?){
         _uiState.update { it.copy(obra=obra) }
